Fixed calculation accuracy on backend

+++ Calculator.Backend.Data/Models/DTO/HistoryDTO.cs
@@ -2,9 +2,9 @@
 {
     public class HistoryDTO
     {
-        public float Param1 { get; set; }
-        public float Param2 { get; set; }
-        public float Result { get; set; }
+        public double Param1 { get; set; }
+        public double Param2 { get; set; }
+        public double Result { get; set; }
         public required string OperationName { get; set; }
     }
 }
