+++ Calculator.Backend.Data/Models/DTO/OperationParamsDTO.cs
@@ -1,13 +1,14 @@
 ﻿using Calculator.Backend.Data.Converters;
 using Calculator.Backend.Data.Enums;
+using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
 namespace Calculator.Backend.Data.Models.DTO
 {
     public class OperationParamsDTO
     {
-        public float Param1 { get; set; }
-        public float Param2 { get; set; }
+        public double Param1 { get; set; }
+        public double Param2 { get; set; }
         [JsonConverter(typeof(OperationTypeConverter))]
         public OperationType OperationType { get; set; }
     }
