+++ Calculator.Backend.Data/Models/Entities/History.cs
@@ -3,9 +3,9 @@
     public class History
     {
         public Guid Id { get; set; }
-        public float Param1 { get; set; }
-        public float Param2 { get; set; }
-        public float Result { get; set; }
+        public double Param1 { get; set; }
+        public double Param2 { get; set; }
+        public double Result { get; set; }
 
         public Guid OperationId { get; set; }
         public Operation Operation { get; set; }
